requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymysql==1.1.0
sqlalchemy==2.0.23
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==1.10.12
email-validator==2.1.0